user-liaison/requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq==0.9.0
httpx==0.27.2
python-dotenv==1.0.0
